split graph.py into separate rdf modules
